pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "RAG system with access-aware retrieval and evaluation"
readme = "README.md"
requires-python = ">=3.10"

authors = [
    { name = "Rizwan" }
]

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "langchain>=0.1",
    "langgraph>=0.0.30",
    "openai>=1.12",
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2",
    "langfuse>=2.20",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "langchain-text-splitters>=1.1.0",
    "prometheus-client>=0.20",
    "streamlit>=1.32",
    "email-validator>=2.1",
    "pypdf>=4.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "Pillow>=10.0",
    "python-multipart>=0.0.22",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "python-pptx>=0.6.23",
    "redis>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "black>=24.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
